able to calculate floating value

+++ app.js
@@ -1,6 +1,6 @@
 function areaPer() {
-  var w = parseInt(document.getElementById("width").value);
-  var h = parseInt(document.getElementById("height").value);
+  var w = parseFloat(document.getElementById("width").value);
+  var h = parseFloat(document.getElementById("height").value);
   if (w < 0 || h < 0) {
     alert("Height or width cannot be negative!");
     return;
